requirements.txt
torch
torchvision
numpy
pandas
Pillow
matplotlib
scikit-learn
seaborn
tqdm
pytesseract
pdf2image
PyMuPDF
